chore: allow posthog to maintain a state of feature flags (#708)
So one solution to the concern around feature flags making an API call
everytime one is checked. If you add a posthog personal API key their
SDK creates a poller to maintain the feature flag state. Not as nice as
some other products with SSE, but it's what they've got. We already have
a personal API key available in our GCP environment so it's just a
matter of plumbing it through.

https://github.com/PostHog/posthog-go/blob/master/posthog.go#L157

The alternative to this I think is we don't do this and we just read
through cache feature flags ourselves