chore: fix feature flagging for logs (#707)
1. We were never using a real feature flags client. conv.Default[...]
was always falling back to the InMemory stub which naturally was empty
on live environments. Best explanation I can give as to why is some
specific on how go interfaces work
```
Because features is technically non-nil, conv.Default thinks there is a valid provider — but since the underlying value is nil, it can’t be used, so it always falls back to &feature.InMemory{}.
```

2. Using the same feature flag on the frontend and backend creates some
weirdness when setting it. Right now the backend works off
organization_id but the frontend works off user identity. So you
basically have to set a condition for the flag twice one for the org,
one for user so they see it on the frontend. I think it is a lot cleaner
just to pass this value from the backend, so setting the flag to true
for an organization_id is always the source of truth as to whether logs
are on.