feat: correctly trim fragments in plan execution (#727)
fixes trimming fragments in plan execution

Fragments are valid to be included in OpenAPI paths for differentiation
but they should not be included in the actual request to the server. I
think trimming this at the Plan Execution level is a decent option since
it ensures request URL is valid as we move into the tool proxy.

Another option would be to trim these at deployment time when we enter
the request URL, but I do think there's some value to having it
represent exactly what was in OpenAPI