fix: allow promises as default export in Gram Functions entrypoints (#730)
This change updates the Gram Functions runner to detect if the default
export from customer code is a Promise to an object containing
`handleToolCall` / `handleResources` and awaits it before proceeding
with a tool/resource request.

We also include some improvements to the FunctionsError class around
capturing the true causing error.