fix: auth tab don't remove unset values (#771)
Currently, if you set a value in the Auth tab of the toolset page it
will erase all other, previously set environment entries. This fixes
that, with the caveat that you now won't be able to unset environment
entries. We'll have to support that another way (though you can still do
it through the dedicated environments page)