chore: lift collection of tool call logs (#777)
The original tool call log collection was written all within a
roundtripper inside of `reverseProxyRequest`. This is a problem because
there are plenty of valid places that a tool call could not make it to
that point.

- A problem with function runner
- Input not passing JSON Schema Validation
- Some OAuth credential exchange issue
- Generally not at least showing that a custom tool was called

This meant the logs we were tracking did not line up with the tool calls
we are tracking for our polar events or business metrics.

Trying to not completely change this. This keeps the roundtripper for
gathering downstream http specifics on request, but allows us still to
lift the actual log collection to a higher level where we currently emit
tracking metrics. There can probably be further joining of this
functionality, but this feels reasonable for now