feat: small ux tweaks to logging page (#784)
- Makes tool type filterable on more than just http tools (functions,
custom)
- Renames the concept of servers to sources (which is accurate)
- This was a concept misnamed from the start. The functionality that is
actually being provided is filtering by source not server. So the naming
needs to reflect that. Filtering by server would be great, but there is
nothing being stored related to the "server" on the tool call log right
now so that would be a feature someone needs to add.
- Makes log details scrollable, it was getting cutoff