feat: redacted security headers in opt-in tool call logs (#788)
- Supports adding refacted security headers that get set for http tools
from http_security in opt in tool call logs
- Creates a relatively simple heuristic for redaction
- Removes the Client column from the logs page. This is never going to
be set, the way it's being tracked right now is not accurate. We'd need
to ensure we pick up client from upstream MCP provider which could be
done at a later time. But right now having the column with no data is a
poor experience