chore: add toolset embeddings (#817)
- entry_key which is prefixed allows us to embed and run separate
semantic searches on entries of multiple types for toolsets. Right now
we are only doing `tool:`
- In pgvector, to have a `hnsw` or `ivfflat` index you MUST define an
explicit dimensionality to your vector column. In this case `1536` for
openai/text-embedding-3-small. On general research prevailing sentiment
seems to be that if you want to support embeddings of multiple different
models either have separate tables or separate columns for the embedding
vector. This column is named so that embedding vector columns of
different sizes could be defined and indexed.
- You can try to pad vectors of smaller size. But it's really not
recommended results wise for pgvector