fix: suppress excessive reaper logging and update reaper query (#832)
This change contains two fixes related to the gram functions reaper.

## fix: suppress logging to user deployment logs

This change modifies the function reaping process to reduce noise in
user deployment logs by suppressing routine informational messages.

## fix: scope app reaper query to project id

This change updates the database query to list reapable fly apps so that
it can be scoped to a specific project ID. This allows project-scoped
reaping. Previously, the project-scoped reaper was not passing the
project ID to the query and it was acting as a global reaper.