feat: product metric attribute estimating tool call success (#919)
The internal request is for a product metric attribute that can estimate
tool call success/failure.

It's important to note this isn't an exact heuristic. But we define one
that has been discussed that we believe is a reasonable estimator of
tool call we know have failed either due to authentication or server
issues.

For now also including status_code directly to let it be sorted out in
posthog. We may eventually settle on just one option