fix: enable unauthenticated running of clickhouse migrations using golang-migrate  (#935)
Based off this
[issue](https://linear.app/speakeasy/issue/AGE-884/discuss-how-hard-is-it-to-remove-the-dependency-on-atlas-for-zero-got)

This PR proposes a solution for the issue of requiring Atlas login in
order to run Clickhouse migrations.

## The Problem

Clickhouse migrations are a paid feature for Atlas. It means that any
user that wants to install/develop gram locally will have to have an
Atlas login, which adds friction (and potential costs).

## The solution

To overcome this, we're generating 2 migration files for Clickhouse when
there's a database change:

1. The normal Atlas migration files
2. Files that can be used by golang-migrate, which is fully free to use

We then have 2 possible paths, controlled by a `mise` variable:

1. Using Atlas for everything, for people that are logged in
2. Using golang-migrate for people that do not wish to log in

The good part about the pattern here is that atlas still derives new
migrations from a main schema (it's still declarative).

## (Mo) Problems

This approach is not pretty:

1. golang-migrate doesn't accept Clickhouse URLs in the format
`user:password@host:port/database` so we had to create a separate
variable for this
2. We have 2 ways of handling migrations. And if a user decides to
change an approach after having run migrations (using golang-migrate
instead of Atlas for example) this will break (I tried locally and
golang-migrate will complain that a table already exists)

## Alternatives

None of them are particularly appealing:
1. We can make Clickhouse be explicit "opt-in". If users do not opt in
we don't send logs to CH (requires code changes to have a noop route for
ingestion)
2. We can keep things as is and require users to login

I still do believe that this approach will be the path of least
friction, but I'm not super happy about it so happy to hear other
potential ideas.