mig: add agent execution table (#961)
A very basic table that allows to track agent execution for the purpose
of visualization/pagination in a dashboard tab.

The goal is not to materialize a bunch of data about the specific
execution. That is currently held in temporal and would be retrieved via
API upon clicking into an agent run.

This will just efficiently power a very basic list view of agent runs by
timestamp. More data could eventually be materialized onto this table,
but for now it should stay very limited.