fix(AddSourceDialog): continue button not closing dialog when clicked (#979)
When adding a source, the continue button didn't "appear" to be doing
anything as the active modal wasn't closing. We route the user to the
toolsets page, but if the user was already on that page, there wasn't a
navigation action or similar to show that we've completed the ask. With
this PR, we close the dialog once it's been successfully submitted and
the user is able to the new toolset on the the toolsets page.