feat: allow uploading OpenAPI specs via remote url (#977)
This change allows users to input a URL which links to an OpenAPI.yaml
file instead of directly uploading a file from the local filesystem. We
fetch the spec contents on the client-side so we can plug in to the
existing spec-upload flow and all of it's existing behaviors.

Note: We're not storing these URLs in this PR but a potential future
upgrade may involve storing the URLs and providing a "sync" option or
similar.