feat: static oauth callback in oauth proxy (#980)
unlike the parts of OAuth that face the MCP client. The oauth redirect
actually does not need to have the MCP slug directly in its URL.

There is a bit advantage to us not doing this. If you have the MCP slug
in the callback URL that means every MCP server in a project needs its
own redirect URL to be whitelisted with the downstream oauth provider
- If a project has multiple servers that need the same oauth they will
need to do a setup for each individual server
- If gram eventually wants to provide oauth integrations out of the box.
It is basically not possible to whitelist a simple
`app.getgram.ai/oauth/callback` for the Gram app

Discussed with Chase. Putting this in state is actually much better. It
allows us to keep this redirect URL stable while still achieving the
same results. Only the endpoints that face the MCP client should have
slugs