fix: inconsistency applying private environment passthrough auth (#997)
In a case where an MCP server is being used as a private server and it
has a default environment attached. If that environment has a certain
variable that's also being passed through directly on use. We should
always prioritize the one that is passed through directly on use.

Prior to this change we were having issues with case insensitivity
`speakeasy_key`(user) `SPEAKEASY_KEY` (env)