fix: ensure react query client is stable between renders
The QueryClient constructed in Sdk.tsx was subject to being recreated
during development mode hot reloads and it was causing unnecessary
query cache invalidations. This change refactors the code so that the
QueryClient is only created once per client-side session. This includes
the error handling logic which it depends on. Error handling functions
did not need to be threaded through React context so that was all
removed in favor of simpler exported functions from `@/lib/errors`.