fix: resolve DialogContent requires a DialogTitle warning (#1008)
Updated the Dialog component used in Sources component to conditionally
render the whole Dialog based on whether the open view is "add-source".
Previously only the contents of the dialog were conditionally rendered,
but with an invalid hierarchy:

```
<Dialog
  open={dialogState.type !== "closed"}
  onOpenChange={(open) => !open && closeDialog()}
>
  <Dialog.Content className="max-w-2xl!">
    {dialogState.type == "add-source" && (
      <AddSourceDialogContent onCompletion={closeDialog} />
    )}
  </Dialog.Content>
</Dialog>
```

In the case above, the dialog can be in a state where `Dialog.Content`
is empty including not containing a `Dialog.Title` child which caused a
flurry of warnings in the console:

```
`DialogContent` requires a `DialogTitle` for the component to be accessible for screen reader users.

If you want to hide the `DialogTitle`, you can wrap it with our VisuallyHidden component.

For more information, see https://radix-ui.com/primitives/docs/components/dialog
```