chore: implement S3 asset store (#1012)
This change implements the asset store interface for S3-compatible
storage backends. This covers support for Tigris which will eventually
be used to power large Gram Functions.

## Additional changes

### Updated `BlobStore.Write`

- No longer takes an io.Reader argument because it was not being used by
anything.
- Now takes a `contentLength` argument which can help various backends
with setting up internal buffers for streaming uploads. This improves
memory usages.

### Added `BlobStore.PresignRead`

Large Gram Functions will require access to assets in the blob store.
This will be done through pre-signed URLs with short lifetimes.
Therefore in this change we introduce and implement this method for
support storage backends.