chore: move location of tool metrics from thirdparty to telemetry folder (#1024)
it felt a bit odd to me that the tool metrics logic lived under the
`thirdparty` folder. As a new-comer it wasn't obvious this lived here,
and this folder does seem dedicated to third party tooling (polar,
slack, etc).

This PR moves it to a dedicated `telemetry` folder. I considered having
it in `o11y` but I felt that folder should be dedicated to the actual
server logging tooling (slog, metrics, etc).

I'm doing this in preparation of work to start fetching function logs
from clickhouse to our dashboard.