chore: add functions service for presigning code assets (#1025)
This change adds a new service to Gram that handles presigning assets
for large Gram Functions. This service will be used by function runners
on flyio to download code assets securely on boot. It uses a distinct
JWT auth mechanism that runners are expected to create and sign using
the shared secret configured in Gram and injected into them at
deployment time.