chore: add queries to fetch tool logs from clickhouse (#1029)
We want to expose logs from Clickhouse to our dashboard (and any users
that might want them). This adds a query to fetch the logs with basic
filtering and pagination.

SQLC doesn't support clickhouse so I tried to keep something consistent
with our SQLC definitions:

- We still write CH queries in a `queries.sql` file
- We can then ask Claude to autogenerate the code for us

It's not ideal, but it's a way to still keep something consistent while
taking advantage of some code generation.

I've added a readme and instructions so Claude can work nicely with it.

Note: the folder structure isn't very consistent with the rest of our
codebase, but I'm going to refactor this (adding `repo` folder and
moving the `impl.go` logic from `log/` to here).