chore: refactor telemetry structure (#1038)
## Why?

It came to my attention while implementing the functionality to add
queries for our tool logs that the telemetry folder structure didn't
follow our conventions. Namely:

1. The repo folder didn't exist - business logic and data access logic
were collapsed in one folder
2. There was a separate `logs` folder that was the main service
entrypoint - which is different than our practice of keeping this next
to the main business logic / repo

## What

This refactors the folder structure to be in line with our conventions.
More importantly, it makes working with the new queries and service
implementation easier and maintainable -as now the whole code for
telemetry logic will live in once place.