feat: support large gram functions (#969)
This change introduces support for large Gram Functions. Previously,
Gram Functions could only be around 700KiB zipped which was adequate for
many use cases but was severely limiting for many others. One example is
ChatGPT Apps which can be full fledged React applications with images,
CSS and JS assets embedded alongside an MCP server and all running in a
Gram Function. Many such apps may not fit into this constrained size.
Large Gram Functions addresses this limitation by allowing larger zip
files to be deployed with the help of Tigris, an S3-compatible object
store that integrates nicely with Fly.io - where we deploy/run Gram
Functions.

During the deployment phase on Gram, we detect if a Gram Function's
assets exceed the size limitation and, instead of attaching them in the
fly.io machine config directly, we upload them to Tigris and mount a
lazy reference to them into machines.

When a machine boots up to serve a tool call (or resource read), it runs
a bootstrap process and detects the lazy file representing the code
asset. It then makes a call to the Gram API to get a pre-signed URL to
the asset from Tigris and downloads it directly from there. Once done,
it continues initialization as normal and handles the tool call.

There is some overhead in this process compared to directly mounting
small functions into machines but for a 1.5MiB file, manually testing
indicated that this is still a very snappy process overall with very
acceptable overhead (<50ms). In future PRs, we'll export measurements so
users can observe this.

<img width="3355" height="3172" alt="image"
src="https://github.com/user-attachments/assets/fca5d11c-61a6-4924-a039-e7e6dcbcf079"
/>

---

## Reviewer notes

Some interesting bits in this PR to pay close attention to:

- `serializeAssets` in `server/internal/functions/deploy_fly.go` and
`resolveLazyFile` in `functions/internal/bootstrap/prepare.go` are
probably the most important changes in this PR
- The runner codebase (functions/) now depends on the Goa-generated API
client in the server (`server/gen`). This allows for a type-safe
interaction between runner and API.
- The runner now makes an API call to the Gram server during its
initialization and authenticates using a JWT bearer. You can see how it
is constructed in `functions/internal/auth/jwt.go`. Being able to
construct a subject with project, deployment and function id and signing
it with the shared secret provides enough assurance that a legitimate
entity is requesting assets from Gram.