chore: add retry mechanism when attempting to connect to clickhouse (#1082)
[Thread for
context](https://speakeasyapi.slack.com/archives/C08H55TP4HZ/p1765817625182539)

Our Clickhouse cloud instance in Dev will go into an idle state after
30s. If the machine is in an idle state when we attempt to connect it
takes a while to wakeup and respond to a ping request, and often times
out.

This PR adds a simple retry mechanism with exp. backoff so that if we
run into this issue we give it some time for the machine to come back
from idle.

We were also returning a stub of clickhouse when an error occurred,
which made it incredibly hard to debug issues related to Clickhouse - so
now we return an error instead.