chore: add table schema and migrations to store function logs (#1081)
This is an evolution of the `tool_logs` table after some discussions
that were had as a team.

There are 2 key changes present in this table:

1. It now more closely resembles what we should expect from [OTel Logs
data model](https://opentelemetry.io/docs/specs/otel/logs/data-model/),
which brings more consistency in how we serve logs. This is especially
important if we are to export them to users so that they can import
these into their own observability backend.
2. Our previous strategy involved storing HTTP requests and function
logs separately. It's becoming very evident that these 2 are closely
related, and users will likely want to query them all. It also saves us
from awkward joins between http requests and function logs (we can just
use group by trace ID to get all logs for a trace for example), and
allows us to build materialized tables on top.

Subsequently we'll also update our API endpoints and responses to match
the OTel specs.

## Noteworthy comments

### Body field

The `body` field  can be many things, depending on the log:

* If the log comes from HTTP request, this won't matter much (we can
store some blurb like "HTTP request successful")

* If the log comes from our server (either a gram function runner or
gram server) we log the `msg|message` field from the log.

Example:
Raw log:

```
{"timestamp": ..., "some_attr": "bla", "msg": "some message" ...}
```

Table entry:
```
body: "some message"
```

* If the log comes from a users' function, we write the raw user log to
the body. We could try and find a `message` field within the user's log,
and just add that. But it feels like it'd make more sense to just add
the full log to the body (OTEL actually mentions that body can be a
structured log)

Example:
Raw log:

```
{"timestamp": ..., "some_attr": "bla", "msg": "log", "payload": "{"attr": "bla", "message": "this is a log"}",...}
```

Table entry:
```
body: "{"attr": "bla", "message": "this is a log"}"
```



### Denormalized fields

Some of the columns of the tabled are denormalized from attributes
fields. This is so that we can query for fundamental parameters (such as
gram project or tool URN).

### Order & Indices

I've made a bit of a judgement call on these indices, as we don't yet
know the real world usage query patterns. Things to keep in mind:

1. Order key is set to project ID and timestamp. We will ALWAYS want to
fetch logs by project ID so this makes sense.
3. While tool URN, function ID and deployment ID will also be common
filtering fields, that won't ALWAYS be the case. Sometimes we'll query
by project and deployment, other times by project and tool. For that
reason these were left as indices rather than order keys.
4. I added a few other indices that are likely hits, but we can always
experiment and tweak.

I'm not overly concerned about storage and performance costs as we only
retain 30days logs anyways. We'll observe and monitor the performance
and adjust as needed.

## Next steps

1. Generate SQL queries to retrieve/insert data into the new table
2. Create new queries to read/write table
3. Update existing code to start writing to new table
4. Update API for log retrieval
5. Update dashboard to include function logs in logs tab

## Reference

For validation/reference, here is an example of the model as
[specified](https://opentelemetry.io/docs/specs/otel/protocol/file-exporter/#examples)
by OTel:

```
{
  "resourceLogs": [
    {
      "resource": {
        "attributes": [
          {
            "key": "resource-attr",
            "value": { "stringValue": "resource-attr-val-1" }
          }
        ]
      },
      "scopeLogs": [
        {
          "scope": {},
          "logRecords": [
            {
              "timeUnixNano": "1581452773000000789",
              "severityNumber": 9,
              "severityText": "Info",
              "body": { "stringValue": "This is a log message" },
              "attributes": [
                { "key": "app", "value": { "stringValue": "server" } },
                { "key": "instance_num", "value": { "intValue": "1" } }
              ],
              "droppedAttributesCount": 1,
              "traceId": "08040201000000000000000000000000",
              "spanId": "0102040800000000"
            },
            {
              "timeUnixNano": "1581452773000000789",
              "severityNumber": 9,
              "severityText": "Info",
              "body": { "stringValue": "something happened" },
              "attributes": [
                { "key": "customer", "value": { "stringValue": "acme" } },
                { "key": "env", "value": { "stringValue": "dev" } }
              ],
              "droppedAttributesCount": 1,
              "traceId": "",
              "spanId": ""
            }
          ]
        }
      ]
    }
  ]
}
```