feat: add more robust discovery of oauth metadata (#1097)
Instead of assuming locations on well-known's at the mcp host, we follow
the proper protocol of reading `www-authenticate` and formulate gram
hosted metadata that proxies to the correct authorization servers.

Previous implementation relied only on a semi-reliable heuristic of
discovering at the host, but will fall over.

In this discovery implementation:
1. Attempt explicit oauth
2. Probe the oauth discovery endpoints of the external MCP endpoint in
the same precedence order as the typescript MCP SDK
3. We persist the external OAuth metadata in a way that can be robustly
mapped to gram authorization metadata

Notably, this change seems to fix Claude Code installs for not
_entirely_ obvious reasons, but it seems to result in all gram servers
returning more accurate OAuth metadata